requirements.txt
numpyopencv-pythonscikit-learntorchtorchvisionmlflowfastapiuvicornpillowpython-multipartpytest